% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{buildDatabase}
\alias{buildDatabase}
\title{buildDatabase}
\usage{
buildDatabase(id, name, schema)
}
\value{
a Database instance
}
\description{
buildDatabase
}
\details{
Creates a Database identified by an id, labeled by a name,
and contained in a scheama. This class is designed to
be used as an element in the database list when calling
buildServer
}

