% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{buildServer}
\alias{buildServer}
\title{buildServer}
\usage{
buildServer(name, dbms, hostname, port, databases)
}
\value{
a Server instance
}
\description{
buildServer
}
\details{
Creates Server which can contain 1 or more database
instances.
}

