% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clean}
\alias{clean}
\title{Clean Tables}
\usage{
clean(serverList, targetDatabaseSchema, tablestem = "pa_")
}
\value{
none
}
\description{
Clean Tables
}
\details{
Removes all tables related to the pregnancy algorithm
}

