% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{execute}
\alias{execute}
\title{Execute Algorithm}
\usage{
execute(serverList, targetDatabaseSchema, tablestem = "pa_")
}
\value{
none
}
\description{
Execute Algorithm
}
\details{
Executes the pregnancy identification algorithm. init() must
be called before executing this.
}

