% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{init}
\alias{init}
\title{Init Tables}
\usage{
init(serverList, targetDatabaseSchema, tablestem = "pa_")
}
\value{
none
}
\description{
Init Tables
}
\details{
Initalizes lookup and result tables for pregnancy algorithm.
}

